+++ workspace/ims-webapp/src/main/java/th/co/baiwa/excise/ia/persistence/repository/TransferListRepository.java
@@ -1,5 +1,6 @@
 package th.co.baiwa.excise.ia.persistence.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.Query;
@@ -7,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.co.baiwa.excise.ia.persistence.entity.TransferList;
 
-public interface TransferListRepository extends CommonJpaCrudRepository<TransferList, Long> {
+public interface TransferListRepository extends CommonJpaCrudRepository<TransferList, BigDecimal> {
 
 	@Query(
 			value = "SELECT * FROM IA_TRANSFER_LIST T WHERE T.IS_DELETED = 'N' AND T.TRANSFER_LIST = ? AND T.BUDGET_TYPE = ? AND T.ACTIVITIES = ?  ORDER BY T.TRANSFER_ID ASC",
@@ -17,6 +18,4 @@ public interface TransferListRepository extends CommonJpaCrudRepository<Transfer
 //	, Date start, Date end
 //	T.IS_DELETED = '"+FLAG.N_FLAG+"' AND
 	
-	
-
 }
